[#28] Time filters for bug + responsive grid
