+++ AnimalPocket/AnimalPocket/Views/Helpers/LowTideBackground.swift
@@ -8,16 +8,16 @@
 import SwiftUI
 
 struct LowTideBackground: View {
-    var body: some View {
-        Image("lowTide")
-        .resizable()
-        .aspectRatio(contentMode: .fill)
-        .ignoresSafeArea()
-    }
+  var body: some View {
+    Image("lowTide")
+      .resizable()
+      .aspectRatio(contentMode: .fill)
+      .ignoresSafeArea()
+  }
 }
 
 struct LowTideBackground_Previews: PreviewProvider {
-    static var previews: some View {
-        LowTideBackground()
-    }
+  static var previews: some View {
+    LowTideBackground()
+  }
 }
